+++ benches/mat3bench.rs
@@ -10,15 +10,15 @@ use criterion::{criterion_group, criterion_main, Criterion};
 
 fn bench_mat3_transpose(c: &mut Criterion) {
     let mut group = c.benchmark_group("mat3 transpose");
-    group.bench_function("glam", |b| {
+    bench_glam!(group, |b| {
         use glam::Mat3;
         bench_unop!(b, op => transpose, ty => Mat3)
     });
-    group.bench_function("cgmath", |b| {
+    bench_cgmath!(group, |b| {
         use cgmath::{prelude::*, Matrix3};
         bench_unop!(b, op => transpose, ty => Matrix3<f32>)
     });
-    group.bench_function("nalgebra", |b| {
+    bench_nalgebra!(group, |b| {
         use nalgebra::Matrix3;
         bench_unop!(b, op => transpose, ty => Matrix3<f32>)
     });
@@ -27,19 +27,19 @@ fn bench_mat3_transpose(c: &mut Criterion) {
 
 fn bench_mat3_determinant(c: &mut Criterion) {
     let mut group = c.benchmark_group("mat3 determinant");
-    group.bench_function("glam", |b| {
+    bench_glam!(group, |b| {
         use glam::Mat3;
         bench_unop!(b, op => determinant, ty => Mat3)
     });
-    group.bench_function("cgmath", |b| {
+    bench_cgmath!(group, |b| {
         use cgmath::{prelude::*, Matrix3};
         bench_unop!(b, op => determinant, ty => Matrix3<f32>)
     });
-    group.bench_function("nalgebra", |b| {
+    bench_nalgebra!(group, |b| {
         use nalgebra::Matrix3;
         bench_unop!(b, op => determinant, ty => Matrix3<f32>)
     });
-    group.bench_function("euclid", |b| {
+    bench_euclid!(group, |b| {
         use euclid::{Transform2D, UnknownUnit};
         bench_unop!(b, op => determinant, ty => Transform2D<f32, UnknownUnit, UnknownUnit>)
     });
@@ -48,19 +48,19 @@ fn bench_mat3_determinant(c: &mut Criterion) {
 
 fn bench_mat3_inverse(c: &mut Criterion) {
     let mut group = c.benchmark_group("mat3 inverse");
-    group.bench_function("glam", |b| {
+    bench_glam!(group, |b| {
         use glam::Mat3;
         bench_unop!(b, op => inverse, ty => Mat3)
     });
-    group.bench_function("cgmath", |b| {
+    bench_cgmath!(group, |b| {
         use cgmath::{prelude::*, Matrix3};
         bench_unop!(b, op => invert, ty => Matrix3<f32>)
     });
-    group.bench_function("nalgebra", |b| {
+    bench_nalgebra!(group, |b| {
         use nalgebra::Matrix3;
         bench_unop!(b, op => try_inverse, ty => Matrix3<f32>)
     });
-    group.bench_function("euclid", |b| {
+    bench_euclid!(group, |b| {
         use euclid::{Transform2D, UnknownUnit};
         bench_unop!(b, op => inverse, ty => Transform2D<f32, UnknownUnit, UnknownUnit>)
     });
@@ -70,19 +70,19 @@ fn bench_mat3_inverse(c: &mut Criterion) {
 fn bench_mat3_mul_mat3(c: &mut Criterion) {
     use std::ops::Mul;
     let mut group = c.benchmark_group("mat3 mul mat3");
-    group.bench_function("glam", |b| {
+    bench_glam!(group, |b| {
         use glam::Mat3;
         bench_binop!(b, op => mul, ty1 => Mat3, ty2 => Mat3)
     });
-    group.bench_function("cgmath", |b| {
+    bench_cgmath!(group, |b| {
         use cgmath::Matrix3;
         bench_binop!(b, op => mul, ty1 => Matrix3<f32>, ty2 => Matrix3<f32>, param => by_ref)
     });
-    group.bench_function("nalgebra", |b| {
+    bench_nalgebra!(group, |b| {
         use nalgebra::Matrix3;
         bench_binop!(b, op => mul, ty1 => Matrix3<f32>, ty2 => Matrix3<f32>, param => by_ref)
     });
-    group.bench_function("euclid", |b| {
+    bench_euclid!(group, |b| {
         use euclid::{Transform2D, UnknownUnit};
         bench_binop!(b, op => post_transform, ty => Transform2D<f32, UnknownUnit, UnknownUnit>, param => by_ref)
     });