+++ benches/quatbench.rs
@@ -5,19 +5,19 @@ use criterion::{criterion_group, criterion_main, Criterion};
 
 fn bench_quat_conjugate(c: &mut Criterion) {
     let mut group = c.benchmark_group("quat conjugate");
-    group.bench_function("glam", |b| {
+    bench_glam!(group, |b| {
         use glam::Quat;
         bench_unop!(b, op => conjugate, ty => Quat)
     });
-    group.bench_function("cgmath", |b| {
+    bench_cgmath!(group, |b| {
         use cgmath::Quaternion;
         bench_unop!(b, op => conjugate, ty => Quaternion<f32>)
     });
-    group.bench_function("nalgebra", |b| {
+    bench_nalgebra!(group, |b| {
         use nalgebra::UnitQuaternion;
         bench_unop!(b, op => conjugate, ty => UnitQuaternion<f32>)
     });
-    group.bench_function("euclid", |b| {
+    bench_euclid!(group, |b| {
         use euclid::{Rotation3D, UnknownUnit};
         // inverse assume normalized quaternion, so it's just a conjugate
         bench_unop!(b, op => inverse, ty => Rotation3D<f32, UnknownUnit, UnknownUnit>)
@@ -28,19 +28,19 @@ fn bench_quat_conjugate(c: &mut Criterion) {
 fn bench_quat_mul_quat(c: &mut Criterion) {
     use std::ops::Mul;
     let mut group = c.benchmark_group("quat mul quat");
-    group.bench_function("glam", |b| {
+    bench_glam!(group, |b| {
         use glam::Quat;
         bench_binop!(b, op => mul, ty1 => Quat, ty2 => Quat)
     });
-    group.bench_function("cgmath", |b| {
+    bench_cgmath!(group, |b| {
         use cgmath::Quaternion;
         bench_binop!(b, op => mul, ty1 => Quaternion<f32>, ty2 => Quaternion<f32>)
     });
-    group.bench_function("nalgebra", |b| {
+    bench_nalgebra!(group, |b| {
         use nalgebra::UnitQuaternion;
         bench_binop!(b, op => mul, ty1 => UnitQuaternion<f32>, ty2 => UnitQuaternion<f32>)
     });
-    group.bench_function("euclid", |b| {
+    bench_euclid!(group, |b| {
         use euclid::{Rotation3D, UnknownUnit};
         bench_binop!(b, op => pre_rotate, ty => Rotation3D<f32, UnknownUnit, UnknownUnit>, param => by_ref)
     });