+++ tests/quat.rs
@@ -1,7 +1,7 @@
 mod support;
 use cgmath;
 use glam;
-use mathbench::*;
+use mathbench::mint_support::{random_mint_quat, random_mint_vec3};
 use nalgebra;
 use rand::SeedableRng;
 use rand_xoshiro::Xoshiro256Plus;